+++ rope/contrib/autoimport/__init__.py
@@ -1,0 +1,2 @@
+from .autoimport import AutoImport
+__all__ = ["AutoImport"]
