+++ rope/contrib/autoimport/autoimport.py
@@ -1,248 +1,20 @@
-import ast
-import inspect
 import pathlib
 import re
 import sqlite3
 import sys
-from collections import OrderedDict
 from concurrent.futures import ProcessPoolExecutor
-from enum import Enum
-from importlib import import_module
-from typing import Iterable, List, Optional, Set, Tuple
+from typing import List, Optional, Tuple
 
 from rope.base import exceptions, libutils, resourceobserver, taskhandle
 from rope.base.project import Project
 from rope.base.resources import Resource
 from rope.refactor import importutils
 
-
-class Source(Enum):
-    PROJECT = 0  # Obviously any project packages come first
-    MANUAL = 1  # Any packages manually added are probably important to the user
-    BUILTIN = 2
-    STANDARD = 3  # We want to favor standard library items
-    SITE_PACKAGE = 4
-    UNKNOWN = 5
-
-
-Name = Tuple[str, str, str, int]
-
-
-def _get_names(
-    modpath: pathlib.Path,
-    modname: str,
-    package_name: str,
-    package_source: Source,
-    underlined: bool = False,
-) -> List[Name]:
-    """Update the cache for global names in `modname` module
-
-    `modname` is the name of a module.
-    """
-    # TODO use __all__ parsing if avalible
-    if modpath.is_dir():
-        names: List[Name]
-        if modpath / "__init__.py":
-            names = _get_names_from_file(
-                modpath / "__init__.py",
-                modname,
-                package_name,
-                package_source,
-                only_all=True,
-            )
-            if len(names) > 0:
-                return names
-        names = []
-        for file in modpath.glob("*.py"):
-            names.extend(
-                _get_names_from_file(
-                    file,
-                    modname + f".{file.name.removesuffix('.py')}",
-                    package_name,
-                    package_source,
-                    underlined=underlined,
-                )
-            )
-        return names
-    else:
-        return _get_names_from_file(
-            modpath, modname, package_name, package_source, underlined=underlined
-        )
-
-
-class PackageType(Enum):
-    STANDARD = 1  # Just a folder
-    COMPILED = 2  # .so module
-    SINGLE_FILE = 3  # a .py file
-
-
-def _get_names_from_builtins(
-    underlined: bool = False, packages: Iterable[str] = sys.builtin_module_names
-) -> List[Name]:
-    """Gets names from builtin modules. These are the only modules it is safe to get the names from"""
-    results: List[Name] = []
-    for package in packages:
-        if package == "builtins":
-            continue  # Builtins is redundant since you don't have to import it.
-        if underlined or not package.startswith("_"):
-            module = import_module(package)
-            if hasattr(module, "__all__"):
-                for name in module.__all__:
-                    results.append((str(name), package, package, Source.BUILTIN.value))
-            for name, value in inspect.getmembers(module):
-                if underlined or not name.startswith("_"):
-                    if inspect.isclass(value) or inspect.isfunction(value) or inspect.isbuiltin(value):
-                        results.append(
-                            (str(name), package, package, Source.BUILTIN.value)
-                        )
-
-    return results
-
-
-def _get_package_name_from_path(
-    package_path: pathlib.Path,
-) -> Optional[Tuple[str, PackageType]]:
-    package_name = package_path.name
-    if package_name.endswith(".egg-info"):
-        return None
-    if package_name.endswith(".so"):
-        name = package_name.split(".")[0]
-        return (name, PackageType.COMPILED)
-        # TODO add so handling
-    if package_name.endswith(".py"):
-        stripped_name = package_name.removesuffix(".py")
-        return (stripped_name, PackageType.SINGLE_FILE)
-    return (package_name, PackageType.STANDARD)
-
-
-def _get_modname_from_path(modpath: pathlib.Path, package_path: pathlib.Path) -> str:
-    package_name: str = package_path.name
-    modname = (
-        modpath.relative_to(package_path)
-        .as_posix()
-        .removesuffix(".py")
-        .replace("/", ".")
-    )
-    modname = package_name if modname == "." else package_name + "." + modname
-    return modname
-
-
-def _find_all_names_in_package(
-    package_path: pathlib.Path,
-    recursive=True,
-    package_source: Source = None,
-    underlined: bool = False,
-) -> List[Name]:
-    package_tuple = _get_package_name_from_path(package_path)
-    if package_tuple is None:
-        return []
-    package_name, package_type = package_tuple
-    if package_source is None:
-        package_source = get_package_source(package_path)
-    modules: List[Tuple[pathlib.Path, str]] = []
-    if package_type is PackageType.SINGLE_FILE:
-        modules.append((package_path, package_name))
-    elif package_type is PackageType.COMPILED:
-        return []
-    elif recursive:
-        for sub in _submodules(package_path):
-            modname = _get_modname_from_path(sub, package_path)
-            if underlined or modname.__contains__("_"):
-                continue  # Exclude private items
-            modules.append((sub, modname))
-    else:
-        modules.append((package_path, package_name))
-    result: List[Name] = []
-    for module in modules:
-        result.extend(
-            _get_names(module[0], module[1], package_name, package_source, underlined)
-        )
-    return result
-
-
-def _get_names_from_file(
-    module: pathlib.Path,
-    modname: str,
-    package: str,
-    package_source: Source,
-    only_all: bool = False,
-    underlined: bool = False,
-) -> List[Name]:
-    with open(module, mode="rb") as file:
-        try:
-            root_node = ast.parse(file.read())
-        except SyntaxError as e:
-            print(e)
-            return []
-    results: List[Name] = []
-    for node in ast.iter_child_nodes(root_node):
-        node_names: List[str] = []
-        if isinstance(node, ast.Assign):
-            for target in node.targets:
-                try:
-                    assert isinstance(target, ast.Name)
-                    if target.id == "__all__":
-                        # TODO add tuple handling
-                        all_results: List[Name] = []
-                        assert isinstance(node.value, ast.List)
-                        for item in node.value.elts:
-                            assert isinstance(item, ast.Constant)
-                            all_results.append(
-                                (
-                                    str(item.value),
-                                    modname,
-                                    package,
-                                    package_source.value,
-                                )
-                            )
-                        return all_results
-                    else:
-                        node_names.append(target.id)
-                except (AttributeError, AssertionError):
-                    # TODO handle tuple assignment
-                    pass
-        elif isinstance(node, (ast.FunctionDef, ast.ClassDef)):
-            node_names = [node.name]
-        for node_name in node_names:
-            if underlined or not node_name.startswith("_"):
-                results.append((node_name, modname, package, package_source.value))
-    if only_all:
-        return []
-    return results
-
-
-def get_package_source(
-    package: pathlib.Path, project: Optional[Project] = None
-) -> Source:
-    """Detect the source of a given package. Rudimentary implementation."""
-    if project is not None and package.as_posix().__contains__(project.address):
-        return Source.PROJECT
-    if package.as_posix().__contains__("site-packages"):
-        return Source.SITE_PACKAGE
-    if package.as_posix().startswith(sys.prefix):
-        return Source.STANDARD
-    else:
-        return Source.UNKNOWN
-
-
-def _sort_and_deduplicate(results: List[Tuple[str, int]]) -> List[str]:
-    if len(results) == 0:
-        return []
-    results.sort(key=lambda y: y[-1])
-    results_sorted = list(zip(*results))[0]
-    return list(OrderedDict.fromkeys(results_sorted))
-
-
-def _sort_and_deduplicate_tuple(
-    results: List[Tuple[str, str, int]]
-) -> List[Tuple[str, str]]:
-    if len(results) == 0:
-        return []
-    results.sort(key=lambda y: y[-1])
-    results_sorted = []
-    for result in results:
-        results_sorted.append(result[:-1])
-    return list(OrderedDict.fromkeys(results_sorted))
+from .defs import Name, Source
+from .parse import (_find_all_names_in_package, _get_names_from_builtins,
+                    _get_names_from_file)
+from .utils import (_get_modname_from_path, _get_package_name_from_path,
+                    _sort_and_deduplicate, _sort_and_deduplicate_tuple)
 
 
 class AutoImport(object):
@@ -555,14 +327,3 @@ class AutoImport(object):
                 if package_tuple[0] == package_name:
                     return package
         return None
-
-
-def _submodules(mod: pathlib.Path) -> Set[pathlib.Path]:
-    """Simple submodule finder that doesn't try to import anything"""
-    result = set()
-    if mod.is_dir() and (mod / "__init__.py").exists():
-        result.add(mod)
-        for child in mod.iterdir():
-            result |= _submodules(child)
-    return result
-    return result
